Refactoring and moving into Packages
