+++ app/src/main/java/com/google/firebase/codelab/friendlychat/activities/IndividualChatActivity.java
@@ -52,8 +52,8 @@ import com.google.android.youtube.player.YouTubeStandalonePlayer;
 import com.google.firebase.auth.FirebaseAuth;
 import com.google.firebase.auth.FirebaseUser;
 import com.google.firebase.codelab.friendlychat.R;
-import com.google.firebase.codelab.friendlychat.fragments.MovieFragment;
-import com.google.firebase.codelab.friendlychat.fragments.TrailerFragment;
+import com.google.firebase.codelab.friendlychat.chatAddons.movie.fragments.MovieFragment;
+import com.google.firebase.codelab.friendlychat.chatAddons.movie.fragments.TrailerFragment;
 import com.google.firebase.codelab.friendlychat.models.FriendlyMessage;
 import com.google.firebase.codelab.friendlychat.utilities.ChatApplication;
 import com.google.firebase.codelab.friendlychat.utilities.CodelabPreferences;
