+++ app/src/main/res/layout/fragment_trailer.xml
@@ -4,7 +4,7 @@
     android:layout_width="wrap_content"
     android:layout_height="wrap_content"
     android:layout_gravity="center"
-    tools:context="com.google.firebase.codelab.friendlychat.fragments.TrailerFragment">
+    tools:context="com.google.firebase.codelab.friendlychat.chatAddons.movie.fragments.TrailerFragment">
 
     <!-- TODO: Update blank fragment layout -->
